+++ packages/test/__mocks__/react-native.ts
@@ -1,0 +1,4 @@
+// ignoring react-native imports in tests, we may change this in the future
+module.exports = {
+  Platform: { OS: 'web' },
+}