+++ packages/utils/__mocks__/expo.ts
@@ -1,0 +1,3 @@
+module.exports = require('@baseapp-frontend/test/__mocks__/expo.ts')
+
+export {}