+++ packages/utils/functions/token/setTokenAsync/index.ts
@@ -2,10 +2,11 @@ import { setItemAsync } from 'expo-secure-store'
 import { CookieAttributes } from 'js-cookie'
 
 import { setCookie } from '../../cookie'
+import { isMobilePlatform } from '../../os'
 
 export const setTokenAsync = async (key: string, value: string, config?: CookieAttributes) => {
   try {
-    if (process.env.EXPO_PUBLIC_PLATFORM === 'mobile') {
+    if (isMobilePlatform()) {
       await setItemAsync(key, value)
     } else {
       setCookie(key, value, config)