+++ packages/utils/types/env.ts
@@ -1,0 +1,5 @@
+export type MobileDefaultEnvironmentVariables = {
+  EXPO_PUBLIC_API_BASE_URL: string
+  EXPO_PUBLIC_RELAY_ENDPOINT: string
+  EXPO_PUBLIC_WS_RELAY_ENDPOINT: string
+}